Supported entities should now be encoded and decoded correctly
Added a method to MetaDecoder and MetaEncoder (decode/encode) to decode/encode a value.
Using this method instead of calling init/encode on the values itself will translate this types corectly to the meta format.
Before this commit, types supported by the translator, that stood in the first place were not handled as supported types. Using the two methods, they should be now.

Upgraded Serialization and Representation to use these methods